requirements.txt
altgraph==0.17.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==7.1.2
colorama==0.4.6
idna==3.10
inquirerpy==0.3.4
macholib==1.16.3
packaging==24.2
pefile==2023.2.7
pfzy==0.3.4
prompt_toolkit==3.0.50
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.8.0
shellingham==1.5.4
typer==0.3.2
urllib3==2.3.0
wcwidth==0.2.13
